dev-cit-3.0-sprint3: Update to increase size of uploaded data in glance in DB. THe 8GB upload was failing probably because of this

+++ features/director-async-image-processor/src/main/java/com/intel/director/async/task/UploadTask.java
@@ -199,12 +199,12 @@ public abstract class UploadTask extends ImageActionAsyncTask {
 			ImageStoreUploadTransferObject imageUploadTransferObject = new ImageStoreUploadTransferObject();
 			ImageStoreUploadResponse imageStoreUploadResponse = imageStoreManager
 					.fetchDetails(null, glanceId);
-			int size = (int) (content.length() / 1024);
+			long size = (long) (content.length() / 1024);
 			double dataSize = (content.length() / 1024);
 			ImageAttributes imgAttrs;
 			String uploadid = null;
 			boolean firstTime = true;
-			int sent = (int) imageStoreUploadResponse.getSent() / 1024;
+			long sent = (long) imageStoreUploadResponse.getSent() / 1024;
 			double dataSent = imageStoreUploadResponse.getSent() / 1024;
 			while (dataSent != dataSize) {
